Revert "refactor: move category utility functions to a separate module"
This reverts commit ee793a5f1c960e5e7d625e9f8cddcafee876e609.
