+++ src/modules/address-book/index.mjs
@@ -1,5 +1,4 @@
 export * from "./category.mjs";
-export * from "./category-utils.mjs";
 export * from "./address-book.mjs";
 export * from "./add-to-category.mjs";
 export * from "./remove-from-category.mjs";
